+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
--  `Range<T>` implements `IClonable`
+- `Range<T>` implements `IClonable`
+- Extension methods to convert between `Range<int>` and `System.Range` (only .NET Standard 2.1 and higher)
 <!--
 - The Difference method on IEnumerable<Range<T>> to get the relative complement of two sequences and also reduce the result
 - The Exclusive method on IEnumerable<Range<T>> to get the symmetric difference of two sequences and also reduce the result